* JsonConfigurationFactory and YamlConfigurationFactory now both extend a BaseConfigurationFactory * did the same for the tests * added line break at the end of the test JSON files
