+++ dropwizard-configuration/src/test/resources/factory-test-empty.json
@@ -1,0 +1,1 @@
+
