+++ dropwizard-configuration/src/test/resources/factory-test-invalid.json
