+++ dropwizard-configuration/src/test/resources/factory-test-malformed-advanced.json
