+++ dropwizard-configuration/src/test/resources/factory-test-typo.json
