+++ dropwizard-configuration/src/test/resources/factory-test-unknown-property.json
