+++ dropwizard-configuration/src/test/resources/factory-test-valid.json
