+++ dropwizard-configuration/src/test/resources/factory-test-wrong-type.json
